Backend/requirements.txt
langchain
langchain_groq
agno
python-dotenv
tavily-python
typing
rich
streamlit
langchain_ollama
langchain_community
termcolor
uuid
langchain_text_splitters
langchain_chroma
langchain_core
pydantic
yfinance
beautifulsoup4
requests
fastapi
uvicorn
